Code refactor.
* Move some utility code into pkg/asset in preparation for future
changes.
