+++ pkg/gcv/validator.go
@@ -16,17 +16,14 @@
 package gcv
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"io/ioutil"
 
 	"github.com/forseti-security/config-validator/pkg/api/validator"
+	asset2 "github.com/forseti-security/config-validator/pkg/asset"
 	"github.com/forseti-security/config-validator/pkg/gcv/cf"
 	"github.com/forseti-security/config-validator/pkg/gcv/configs"
 	"github.com/golang/glog"
-	"github.com/golang/protobuf/jsonpb"
-	"github.com/hashicorp/go-multierror"
 	"github.com/pkg/errors"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -179,10 +176,10 @@ func NewValidator(options ...Option) (*Validator, error) {
 // AddData adds GCP resource metadata to be audited later.
 func (v *Validator) AddData(request *validator.AddDataRequest) error {
 	for i, asset := range request.Assets {
-		if err := validateAsset(asset); err != nil {
+		if err := asset2.ValidateAsset(asset); err != nil {
 			return status.Error(codes.InvalidArgument, errors.Wrapf(err, "index %d", i).Error())
 		}
-		f, err := convertResourceViaJSONToInterface(asset)
+		f, err := asset2.ConvertResourceViaJSONToInterface(asset)
 		if err != nil {
 			return status.Error(codes.Internal, errors.Wrapf(err, "index %d", i).Error())
 		}
@@ -190,46 +187,6 @@ func (v *Validator) AddData(request *validator.AddDataRequest) error {
 	}
 
 	return nil
-}
-
-func validateAsset(asset *validator.Asset) error {
-	var result *multierror.Error
-	if asset.GetName() == "" {
-		result = multierror.Append(result, errors.New("missing asset name"))
-	}
-	if asset.GetAncestryPath() == "" {
-		result = multierror.Append(result, errors.Errorf("asset %q missing ancestry path", asset.GetName()))
-	}
-	if asset.GetAssetType() == "" {
-		result = multierror.Append(result, errors.Errorf("asset %q missing type", asset.GetName()))
-	}
-	if asset.GetResource() == nil && asset.GetIamPolicy() == nil {
-		result = multierror.Append(result, errors.Errorf("asset %q missing both resource and IAM policy", asset.GetName()))
-	}
-	return result.ErrorOrNil()
-}
-
-func convertResourceViaJSONToInterface(asset *validator.Asset) (interface{}, error) {
-	if asset == nil {
-		return nil, nil
-	}
-	m := &jsonpb.Marshaler{
-		OrigName: true,
-	}
-	if asset.Resource != nil {
-		cleanStructValue(asset.Resource.Data)
-	}
-	glog.V(logRequestsVerboseLevel).Infof("converting asset to golang interface: %v", asset)
-	var buf bytes.Buffer
-	if err := m.Marshal(&buf, asset); err != nil {
-		return nil, errors.Wrapf(err, "marshalling to json with asset %s: %v", asset.Name, asset)
-	}
-	var f interface{}
-	err := json.Unmarshal(buf.Bytes(), &f)
-	if err != nil {
-		return nil, errors.Wrapf(err, "marshalling from json with asset %s: %v", asset.Name, asset)
-	}
-	return f, nil
 }
 
 // Reset clears previously added data from the underlying query evaluation engine.
